int change to double in determ.cpp; add new test t8.test

+++ determ.cpp
@@ -9,7 +9,7 @@ int main()
 	int mSize = 0;
 	std::cin >> mSize;
 
-	math::Matrix<int> m(mSize, mSize);
+	math::Matrix<double> m(mSize, mSize);
 	for(int i = 0; i < mSize; ++i)
 		for(int j = 0; j < mSize; ++j)
 			std::cin >> m[i][j];
